+++ src/main/java/com/portfolio/beportfolio/repository/StateRepository.java
@@ -5,6 +5,6 @@ import com.portfolio.beportfolio.model.State;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface StateRepository extends JpaRepository<State, Integer>{
+public interface StateRepository extends JpaRepository<State, Long>{
     
 }
